+++ src/dist_comm_vis/adapter/dependency_injection/container.py
@@ -2,7 +2,7 @@ from logging.config import fileConfig
 
 from dependency_injector import containers, providers
 
-from dist_comm_vis.adapter.service.ModelWriterService import JsonModelWriterService
+from dist_comm_vis.adapter.service.JsonModelWriterService import JsonModelWriterService
 from dist_comm_vis.adapter.service.LocalFileFinderService import LocalFileFinderService
 from dist_comm_vis.adapter.service.LocalFileReaderService import LocalFileReaderService
 from dist_comm_vis.adapter.service.LocalFileWriterService import LocalFileWriterService
